% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common_support.R
\name{check_common_support}
\alias{check_common_support}
\title{Check for common support via the standard deviation and chi-square rules}
\usage{
check_common_support(.model)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}
}
\value{

}
\description{
Check for common support via the standard deviation and chi-square rules
}
\author{
Vince Dorrie, George Perrett, Joe Marlo
}
