% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_auto_convert_logicals.R
\name{clean_auto_convert_logicals}
\alias{clean_auto_convert_logicals}
\title{Convert all psuedo-logical columns in a dataframe to booleans}
\usage{
clean_auto_convert_logicals(input_data)
}
\arguments{
\item{input_data}{dataframe}
}
\value{
dataframe
}
\description{
Converts columns of a dataframe containing binary c(0, 1), c("T", "F"), c("True", "False") to boolean c(TRUE, FALSE). Is agnostic to case.
}
\examples{
x <- data.frame(
  zero_one = sample(0:1, 10, replace = TRUE),
  TF = sample(c("T", "F"), 10, replace = TRUE),
  truefalse = sample(c('true', 'false'), 10, replace = TRUE),
  char = sample(LETTERS, 10),
  yn = sample(c("yes", "no"), 10, replace = TRUE)
  )
clean_auto_convert_logicals(x)
}
\author{
Joe Marlo
}
