% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean.R
\name{clean_detect_ZYX_columns}
\alias{clean_detect_ZYX_columns}
\title{Attempt to detect which columns of a dataframe are Z, Y, and X}
\usage{
clean_detect_ZYX_columns(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a list denoting which column names are Z, Y, X, and ID
}
\description{
Attempts to auto detect which columns are treatment, response, or ID columns. Treatment and response are detected based on column name alone; ID is based on column name and the values. If none are detected, then columns are categorized as 'X'. Will only return one item (the first) for Z, Y, and ID.
}
\examples{
.data <- data.frame(
 treatment = c(TRUE, TRUE, FALSE, TRUE, FALSE),
 rsp = c(0.808, 0.296,-1.579,-1.272, 0.627),
 ID = c(0.808, 0.296,-1.579,-1.272, 0.627),
 dummyID = 1:5,
 dummy1 = c(34, 35, 10, 5, 38)
)
clean_detect_ZYX_columns(.data)
}
\author{
Joe Marlo
}
