% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dummies_to_categorical.R
\name{clean_dummies_to_categorical}
\alias{clean_dummies_to_categorical}
\title{Convert dummies in a dataframe to a categorical variable
For downloaded reproducible script use}
\usage{
clean_dummies_to_categorical(df, group_names)
}
\arguments{
\item{df}{dataframe}

\item{group_names}{a vector of strings containing dummies' names of a categorical variable. 
The name of the vector will used as the categorical variable's name}
}
\value{

}
\description{
Convert dummies in a dataframe to a categorical variable
For downloaded reproducible script use
}
\examples{
df <- data.frame(
momwhite = c(1,0,0,0,0,0,1),
momblack = c(0,1,1,1,1,0,0),
momhisp = c(0,0,0,0,0,0,0),
y.obs = rnorm(7, 10, 3)
)
race <- c("momwhite", "momblack", "momhisp")
clean_dummies_to_categorical(df = df, group_names = race)
}
\author{
Junhui Yang
}
