% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dummies_to_categorical.R
\name{clean_dummies_to_categorical_internal}
\alias{clean_dummies_to_categorical_internal}
\title{Convert dummies in the same group to a categorical variable in a dataframe 
For thinkCausal app development use}
\usage{
clean_dummies_to_categorical_internal(
  i,
  df,
  group_names,
  rename_group,
  problematic_group_names
)
}
\arguments{
\item{i}{an integer to indicate the ith group of dummies. Useful to keep track of each group when there are multiple categorical variables in a dataframe}

\item{df}{dataframe}

\item{group_names}{a vector of strings that are the names of dummies in the same group}

\item{rename_group}{a string of the name for the categorical variable}

\item{problematic_group_names}{a vector of strings to indicate which groups have empty names. Initialize it as an empty vector, c(), 
and then the function will return the group number if `rename_group` is an empty string and NUll if `rename_group` is not an empty string}
}
\value{
a list containing group number of problematic group names, a cleaned daatframe with dummies collapsed to a categorical variable, 
and a list of categorical variables' names and their components
}
\description{
Convert dummies in the same group to a categorical variable in a dataframe 
For thinkCausal app development use
}
\examples{
df <- data.frame(
momwhite = c(1,0,0,0,0,0,1),
momblack = c(0,1,1,1,1,0,0),
momhisp = c(0,0,0,0,0,0,0),
y.obs = rnorm(7, 10, 3)
)
clean_dummies_to_categorical_internal(i = 1, df = df, 
                                      group_names = c('momwhite', 'momblack', 'momhisp'), 
                                      rename_group = 'race', problematic_group_names = c())

clean_dummies_to_categorical_internal(i = 1, df = df, 
                                      group_names = c('momwhite', 'momblack', 'momhisp'), 
                                      rename_group = '', problematic_group_names = c())
}
\author{
Junhui Yang
}
