% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean a string for use in column names}
\usage{
clean_names(.names)
}
\arguments{
\item{.names}{}
}
\value{
character vector
}
\description{
Cleans up strings for use as column names in dataframes. Makes the following changes: 
\itemize{
 \item replaces non-ASCII characters with ASCII counterparts  
 \item replaces spaces with underscores  
 \item replaces percent sign with '_percent'  
 \item removes all punctuation except underscores and periods  
 \item adds 'n' to the beginning of strings that start with numeric characters  
}
}
\examples{
.names <- c("yes", "TRUE", "nope\%", "98", 'Ábcdêãçoàúü', 'yep_-,.yep', 'hello goodbye', '', 'no', 'no')
clean_names(.names)
}
\author{
Joe Marlo
}
