% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean.R
\name{clean_to_indicator}
\alias{clean_to_indicator}
\title{Convert categorical columns to indicator variables}
\usage{
clean_to_indicator(df)
}
\arguments{
\item{df}{a dataframe with varibales}
}
\value{
a dataframe with no categorical variables
}
\description{
Convert categorical columns to indicator variables
}
\author{
George Perrett & Joe Marlo
}
