% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_types.R
\name{convert_data_types}
\alias{convert_data_types}
\title{Convert the column types of a dataframe}
\usage{
convert_data_types(.data, new_data_types)
}
\arguments{
\item{.data}{}

\item{new_data_types}{}
}
\value{

}
\description{
Designed to fail in an appropriate way
}
\examples{
x <- data.frame(
  zero_one = sample(0:1, 10, replace = TRUE),
  integers = sample(0:10, 10, replace = TRUE),
  runif = runif(10),
  TF = sample(c("T", "F"), 10, replace = TRUE),
  char = sample(LETTERS, 10)
 )
 str(convert_data_types(x, c('Binary', 'Categorical', 'Continuous', 'Binary', 'Continuous')))
}
