% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_summary_grid.R
\name{create_data_summary_grid}
\alias{create_data_summary_grid}
\title{Create a UI displaying the role, variable name, variable type, and percent NA for each column in a dataframe}
\usage{
create_data_summary_grid(
  .data,
  default_data_types,
  ns_prefix,
  design,
  blocking_variables = NULL
)
}
\arguments{
\item{.data}{a dataframe}

\item{default_data_types}{a vector of default data types. Usually from convert_data_type_to_simple()}

\item{ns_prefix}{a string denoting a prefix to use when creating inputIds}
}
\value{
HTML code
}
\description{
Create a UI displaying the role, variable name, variable type, and percent NA for each column in a dataframe
}
\examples{
# shiny server
# renderUI({
#   create_data_summary_grid(
#     .data = .data,
#     default_data_types = default_data_types,
#     ns_prefix = 'analysis_data_'
#   )
# })
}
