% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_help.R
\name{create_info_icon}
\alias{create_info_icon}
\title{Create hover-able information icon}
\usage{
create_info_icon(label, text)
}
\arguments{
\item{label}{any text to display next to the icon}

\item{text}{text to display when hovering over the icon}
}
\value{

}
\description{
Returns HTML code to create a little (i) icon that shows text when you hover.
}
\examples{
# within shiny UI pages
create_info_icon('Average Treatment Effect', 'The ATE is ...')
}
\author{
Joe Marlo
}
