% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_learning_module.R
\name{create_learning_module}
\alias{create_learning_module}
\title{Build the scaffolding for a new learning module}
\usage{
create_learning_module(
  article_name = "Lorem ipsum",
  include_quiz = TRUE,
  short_name,
  directory_name
)
}
\arguments{
\item{article_name}{The title of the article}

\item{include_quiz}{TRUE/FALSE. Quiz is embedded by default but can be adjusted within the created _module.R file.}

\item{short_name}{A short, unique name for the article to be used for R functions. If not provided, then created from `article_name`.}

\item{directory_name}{The name of the new directory to be created. If not provided, then created from `article_name`.}
}
\value{

}
\description{
Creates the a new learning module from the /modules/learning/_template/ and reminds you of the steps to take to update the app to make it integrate in thinkCausal.
}
\examples{
\dontrun{
create_learning_module('My new article')
}
}
