% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_script}
\alias{create_script}
\title{Build a reproducible script which mimics the thinkCausal work-flow}
\usage{
create_script(
  uploaded_file_name,
  uploaded_file_type,
  uploaded_file_header,
  uploaded_file_delim,
  selected_columns,
  column_names,
  change_data_type,
  estimand,
  common_support
)
}
\value{
string
}
\description{
Build a reproducible script which mimics the thinkCausal work-flow
}
\examples{
create_script(
 uploaded_file_name = 'test.csv',
 uploaded_file_type = 'csv',
 uploaded_file_header = 'TRUE',
 uploaded_file_delim = ',',
 selected_columns = c("Z", "Y", "X1", 'X2', "X3"),
 column_names = c("treatment", "response", "covariate1", "covariate2", "covariate3"),
 estimand = 'att',
 common_support = 'none'
)
}
