% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactive_table.R
\name{create_table}
\alias{create_table}
\title{Automatically create a JavaScript table that has editable cells and checks for the correct input}
\usage{
create_table(
  .data = NULL,
  correct_answers = NULL,
  n_rows = 6,
  y_min = 50,
  y_max = 100,
  ate = 10,
  po_question = TRUE,
  ite_question = TRUE,
  extra_header = NULL,
  extra_header_widths = rep(1, length(extra_header)),
  table_id = NULL,
  ns = NULL
)
}
\arguments{
\item{.data}{optional. A dataframe to generate an interactive table. Any cells that should require user input should be marked with a "?"}

\item{correct_answers}{required only if .data is provided. A vector of the correct answers that map to the "?" rowwise (left-to-right then top-to-bottom)}

\item{n_rows}{an integer to use as the number of rows for the table}

\item{y_min}{a numeric value to use as the lower bound of potential outcomes}

\item{y_max}{a numeric value to use as the upper bound of potential outcomes}

\item{ate}{a numeric value to use as an indicator of positive/negative and magnitude of ITEs}

\item{po_question}{logical indicating if the questions (the places of '?') are at the columns of potential outcomes Y0 and Y1}

\item{ite_question}{logical indicating if the questions (the places of '?') are at the column of ITE}

\item{extra_header}{a vector of column names. If .data is provided, optional to specify; if .data is not provided, keep it equal NULL.}

\item{extra_header_widths}{required only if extra_header is provided. A vector of column widths that sum to length(extra_header). I.e. 1 = width of one column; 2 = width of two columns, etc.}

\item{table_id}{optional. A string to use as the html ID for the table. If none is provided, then 'table' + random 15 letters is used}

\item{ns}{optional. A function defining the name space. To be used with shiny::NS}
}
\value{
a string of JS and HTML code
}
\description{
Automatically create a JavaScript table from a dataframe with editable cells where dataframe=='?'. User must input correct answer into those editable cells. Often used to create potential outcomes quizzes.
}
\examples{
html_string <- create_table(n_rows = 10, y_min = 50, y_max = 100, ate = -10, po_question = T, ite_question = T)
# HTML(html_string) within Shiny UI 
# access the user inputs via get_table_values(<table_id>) within Shiny server
}
\author{
Joe Marlo, Junhui Yang
}
