% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bart.R
\name{fit_bart}
\alias{fit_bart}
\title{Fit a BART model using a standardized interface}
\usage{
fit_bart(.data, support, ran.eff, .estimand)
}
\arguments{
\item{.data}{the data. First column should be treatment, second column response.}

\item{support}{one of c('ate', 'atc', 'att'). See bartCause::bartc}

\item{ran.eff}{if not "None", then the name of the column within .data that represents the random effects}

\item{.estimand}{the causal estimand. See bartCause::bartc}
}
\value{
an object of class "bartcFit"
}
\description{
Fit a BART model using a standardized interface
}
