% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dummies_to_categorical.R
\name{identify_indicators}
\alias{identify_indicators}
\title{Identify levels of an indicator variable when data is has alredy been one hot encoded 
For thinkCausal app development use}
\usage{
identify_indicators(x, cats)
}
\arguments{
\item{x}{a string that corresponds to the name of a variable that may be part of a categorical variable}

\item{cat_df}{dataframe of all non continuous confounders}
}
\value{
a list that containing a vecotr best and a list possible. best is the most likely combination of levels while possible is all possible combinations of levels
}
\description{
Identify levels of an indicator variable when data is has alredy been one hot encoded 
For thinkCausal app development use
}
\examples{
cat_df <- data.frame(
 momwhite = c(1,0,0,0,0,0,1),
 momblack = c(0,1,1,1,1,0,0),
 momhisp = c(0,0,0,0,0,0,0),
 arc =c(1,0,0,0,1,0,1)
 )
identify_indicators(x = 'momwhite', cats = cat_df)
}
\author{
George Perrett
}
