% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_design.R
\name{mod_analysis_design_server}
\alias{mod_analysis_design_server}
\title{analysis_design Server Functions}
\usage{
mod_analysis_design_server(id, store)
}
\description{
analysis_design Server Functions
}
