% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploration.R
\name{plot_exploration}
\alias{plot_exploration}
\title{General purpose plotting for EDA}
\usage{
plot_exploration(
  .data,
  .plot_type = c("Pairs", "Scatter", "Histogram", "Barplot", "Density", "Boxplot"),
  .x,
  .y,
  .fill,
  .fill_static = "grey20",
  .size,
  .shape,
  .alpha = 0.9,
  .vars_pairs,
  .n_bins,
  .jitter,
  .groups,
  .facet,
  .facet_second,
  .include_regression = c("Include", "None")
)
}
\arguments{
\item{.data}{typically store$selected_df}

\item{.plot_type}{one of c("Pairs", 'Scatter', 'Histogram', 'Density', 'Boxplot')}

\item{.x}{}

\item{.y}{}

\item{.fill}{}

\item{.fill_static}{}

\item{.size}{}

\item{.alpha}{}

\item{.vars_pairs}{}

\item{.n_bins}{}

\item{.jitter}{}

\item{.groups}{}

\item{.facet}{}

\item{.facet_second}{}

\item{.include_regression}{one of c("Include", "None")}
}
\value{

}
\description{
To be used programmatically. Not recommended for script/console use.
}
\examples{
library(arm)
data('lalonde', package = 'arm')
X <- lalonde
X <- dplyr::select(X, 'treat', 're78', 'age', 'educ', 'black', 'hisp', 'married', 'nodegr')
X <- clean_auto_convert_logicals(X)
plot_exploration(
 .data = X,
 .plot_type = 'Boxplot', #c("Pairs", 'Scatter', 'Histogram', 'Density', 'Boxplot'),
 .x = 're78',
 .y = 'age',
 .fill = NULL,
 .fill_static = "#5c5980",
 .size = 'age',
 .shape = 'None',
 .alpha = 0.5,
 .vars_pairs,
 .n_bins = 30,
 .jitter = FALSE,
 .groups = 'None',
 .facet = 'None',
 .facet_second = 'None',
 .include_regression = 'None'
)
}
\seealso{
\code{\link{clean_detect_plot_vars}}
}
\author{
Joe Marlo
}
