% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_support_diagnostic.R
\name{plot_tree_diagnostic}
\alias{plot_tree_diagnostic}
\title{Single Regression Tree for common support diagnostic}
\usage{
plot_tree_diagnostic(.model, depth = 2)
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}

\item{confounders}{matrix of confounders}
}
\value{
a list containing variable importance plot & ordered table of confounders by scaled importance
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & table.
}
\author{
George Perrett
}
