% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_message.R
\name{show_message_updating}
\alias{show_message_updating}
\title{Add an blurry overlay and message to user}
\usage{
show_message_updating(div)
}
\arguments{
\item{div}{div ID of the parent}
}
\value{

}
\description{
Blur out an element and add a loading message. This makes it obvious to the user that an element is updating.
}
\examples{
\dontrun{
# shiny server
output$analysis_plot_overlap_plot <- renderPlot({
  
  # add overlay
  show_message_updating('analysis_plot_overlap_plot')
  
  # build plot
  p <- overlap_plot()
  
  # remove overlay
  close_message_updating('analysis_plot_overlap_plot')
  
  return(p)
})
}
}
