% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_settings.R
\name{theme_minimal_no_transparency}
\alias{theme_minimal_no_transparency}
\title{ggplot2 minimal theme without transparent background}
\usage{
theme_minimal_no_transparency(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{See ggplot2::theme_bw}

\item{base_family}{See ggplot2::theme_bw}

\item{base_line_size}{See ggplot2::theme_bw}

\item{base_rect_size}{See ggplot2::theme_bw}
}
\value{

}
\description{
This mimics ggplot2::theme_minimal but excludes the transparent background so downloaded plots are legible.
}
\author{
Joe Marlo
}
