% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_auto_convert_integers.R
\name{clean_auto_convert_integers}
\alias{clean_auto_convert_integers}
\title{Convert integer-like columns with few levels to a factor}
\usage{
clean_auto_convert_integers(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{

}
\description{
Useful for plotting
}
\examples{
x <- data.frame(
  zero_one = sample(0:1, 10, replace = TRUE),
  integers = sample(0:10, 10, replace = TRUE),
  runif = runif(10),
  TF = sample(c("T", "F"), 10, replace = TRUE),
  char = sample(LETTERS, 10)
 )
 str(clean_auto_convert_integers(x))
}
